requirements.txt
# Django Core
Django==5.2
django-filter==24.1
django-extensions==3.2.3

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
drf-spectacular==0.27.0
drf-standardized-errors==0.13.0

# FastAPI & ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
httpx==0.25.2
python-multipart==0.0.6

# Database & ORM
psycopg2-binary==2.9.9
dj-database-url==2.2.0
django-environ==0.11.2
sqlparse==0.5.0

# Real-time & Async
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
aioredis==2.0.1
redis==5.0.7

# Authentication & Security
djangorestframework-jwt==1.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
django-cors-headers==4.3.1
cryptography==42.0.8
PyJWT==2.8.1

# Payment Integration
monnify==0.0.2

# AWS & Cloud Storage
boto3==1.34.129
django-storages[s3]==1.14.2
google-cloud-storage==2.14.0
google-cloud-firestore==2.16.1
firebase-admin==6.1.0

# Phone & Email
django-phonenumber-field==7.3.0
phonenumbers==8.13.38
django-ses==4.1.0

# Frontend & Static Files
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
django-compressor==4.4
whitenoise==6.7.0
django-browser-reload==1.13.0

# Utilities
python-decouple==3.8
decouple==0.0.7
requests==2.32.3
pillow==10.4.0
arrow==1.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
python-slugify==8.0.4

# Monitoring & Logging
sentry-sdk==1.40.6
python-json-logger==2.0.7
django-structlog==7.1.0
structlog==24.1.0
celery==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Admin Enhancements
django-admin-interface==0.25.5
django-import-export==3.3.6
django-tables2==2.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.2.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7

# Production Server
gunicorn==22.0.0
gevent==23.9.1

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.20.0

# Documentation
drf-yasg==1.21.7
sphinx==7.2.6

# Additional utilities
pydantic==2.5.3
pydantic-settings==2.1.0
